Remove SmoothJazz from stats and calculate manually
