+++ packages/farming-weight/src/pests/pest-farming-rate-calculator.ts
@@ -1,5 +1,7 @@
 import { Crop } from '../constants/crops.js';
 import { Pest, SPRAY_TO_PESTS } from '../constants/pests.js';
+import { FarmingPets } from '../constants/pets.js';
+import { Rarity } from '../constants/reforges.js';
 import { Stat } from '../constants/stats.js';
 import type { FortuneUpgrade } from '../constants/upgrades.js';
 import { resolveDropEffects } from '../effects/resolver.js';
@@ -462,8 +464,21 @@ export class PestFarmingRateCalculator {
 
 	private getSmoothJazzMultiplier(): number {
 		const spawnPet = this.player.spawn.selectedPet;
-		const smoothJazzValue = 1 + ((spawnPet !== undefined) ? spawnPet.getFortune(Stat.SmoothJazz) ?? 0 : 0);
-		return smoothJazzValue;
+		if (spawnPet?.type === FarmingPets.Mosquito) {
+			const perLevel = (() => {
+				switch (spawnPet.rarity) {
+					case Rarity.Common:
+					case Rarity.Uncommon:
+						return 0.0025;
+					case Rarity.Rare:
+						return 0.0035;
+					default:
+						return 0.005;
+				}
+			})();
+			return 1 + (spawnPet.level ?? 1) * perLevel;
+		}
+		return 1;
 	}
 
 	private calculatePestDrops(
@@ -874,13 +889,15 @@ function diffValuation(
 	};
 }
 
-function getPestTypeWeights(attraction?: PestAttractionSettings, smoothJazzMultiplier?: number): Partial<Record<Pest, number>> {
+function getPestTypeWeights(
+	attraction?: PestAttractionSettings,
+	smoothJazzMultiplier?: number
+): Partial<Record<Pest, number>> {
 	const pests = attraction?.includeSpecialPests ? [...NATURAL_PESTS, Pest.Mouse, Pest.LunarMoth] : NATURAL_PESTS;
 	const weights: Partial<Record<Pest, number>> = Object.fromEntries(pests.map((pest) => [pest, 1]));
 	if (attraction?.sprayonatorMaterial && SPRAY_TO_PESTS[attraction.sprayonatorMaterial] !== undefined) {
 		for (const spray_pest of SPRAY_TO_PESTS[attraction.sprayonatorMaterial] ?? []) {
-			if (weights[spray_pest] !== undefined)
-				weights[spray_pest] = weights[spray_pest] * 12;
+			if (weights[spray_pest] !== undefined) weights[spray_pest] = weights[spray_pest] * 12;
 		}
 	}
 	if (attraction?.hooveriusVinylTarget && weights[attraction.hooveriusVinylTarget] !== undefined) {
